chore: CI refactor, dependency cleanup, and code extraction
CI / workflows:
- Merge lint.yml + test.yml into a single ci.yml (two parallel jobs)
- Collapse validate-bot / validate-migrations / validate-web into one
  matrix job driven by a compute-matrix step; remove stale
  validate-migrations job (migrations now ship in the bot image)
- Remove docker-pr-build.yml (local testing used instead)
- Remove .flake8 (superseded by ruff)
- Add .coderabbit.yaml with path-specific review instructions derived
  from CLAUDE.md gotchas

Code extraction:
- Extract handle_valkey_command + valkey_listener_loop into
  NerdyPy/utils/valkey.py (pipeline lpush+expire into one round trip)
- Extract parse_config / parse_env_config / deep_merge into
  NerdyPy/utils/config.py

Security / dependencies:
- Migrate web auth from python-jose (CVE-2024-23342, abandoned) to PyJWT
- Replace ffmpeg-python (abandoned since 2019) with subprocess.Popen;
  add returncode check and use context manager for cleanup
- Run uv lock --upgrade (cachetools, filelock, google-auth, virtualenv
  bumped; rsa and future removed as transitive deps)
- Suppress InsecureKeyLengthWarning from PyJWT in pytest config
