--- .flake8
@@ -1,6 +1,0 @@
-[flake8]
-ignore = E203, E266, E501, W503, E722
-max-complexity = 18
-max-line-length = 120
-count = True
-statistics = True
