+++ .github/workflows/ci.yml
@@ -1,10 +1,29 @@
-name: Tests
+name: CI
 
 on:
   pull_request:
     branches: [main]
 
 jobs:
+  lint:
+    name: Ruff Lint & Format Check
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+
+    steps:
+      - uses: actions/checkout@v6
+
+      - name: Run Ruff linter
+        uses: astral-sh/ruff-action@v3
+        with:
+          args: "check --output-format=github"
+
+      - name: Run Ruff formatter check
+        uses: astral-sh/ruff-action@v3
+        with:
+          args: "format --check --diff"
+
   test:
     name: Run Tests
     runs-on: ubuntu-latest
