+++ .github/workflows/docker-pr.yml
@@ -22,7 +22,6 @@ jobs:
       contents: read
     outputs:
       bot: ${{ steps.filter.outputs.bot }}
-      migrations: ${{ steps.filter.outputs.migrations }}
       web: ${{ steps.filter.outputs.web }}
     steps:
       - name: Checkout code
@@ -39,12 +38,6 @@ jobs:
               - 'pyproject.toml'
               - 'uv.lock'
               - 'NerdyPy/**'
-              - '.github/workflows/docker-pr.yml'
-            migrations:
-              - 'Dockerfile'
-              - '.dockerignore'
-              - 'pyproject.toml'
-              - 'uv.lock'
               - 'database-migrations/**'
               - 'alembic.ini'
               - '.github/workflows/docker-pr.yml'
@@ -57,13 +50,41 @@ jobs:
               - 'NerdyPy/**'
               - '.github/workflows/docker-pr.yml'
 
-  validate-bot:
-    name: Validate NerpyBot Image
+  compute-matrix:
+    name: Compute Build Matrix
     runs-on: ubuntu-latest
     needs: [changes]
-    if: needs.changes.outputs.bot == 'true'
+    outputs:
+      matrix: ${{ steps.set.outputs.matrix }}
+      has_targets: ${{ steps.set.outputs.has_targets }}
+    steps:
+      - name: Build target matrix
+        id: set
+        env:
+          BOT: ${{ needs.changes.outputs.bot }}
+          WEB: ${{ needs.changes.outputs.web }}
+        run: |
+          targets=()
+          [[ "$BOT" == "true" ]] && targets+=('{"target":"bot","name":"NerpyBot","tag":"nerpybot:smoke-test"}')
+          [[ "$WEB" == "true" ]] && targets+=('{"target":"web","name":"Web Dashboard","tag":"nerpybot-web:smoke-test"}')
+          if [[ ${#targets[@]} -gt 0 ]]; then
+            matrix=$(printf '%s\n' "${targets[@]}" | jq -sc '{include: .}')
+            echo "has_targets=true" >> "$GITHUB_OUTPUT"
+          else
+            matrix='{"include":[]}'
+            echo "has_targets=false" >> "$GITHUB_OUTPUT"
+          fi
+          echo "matrix=$matrix" >> "$GITHUB_OUTPUT"
+
+  validate:
+    name: Validate ${{ matrix.name }} Image
+    runs-on: ubuntu-latest
+    needs: [compute-matrix]
+    if: needs.compute-matrix.outputs.has_targets == 'true'
     permissions:
       contents: read
+    strategy:
+      matrix: ${{ fromJson(needs.compute-matrix.outputs.matrix) }}
 
     steps:
       - name: Checkout code
@@ -74,79 +95,23 @@ jobs:
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v4
 
-      - name: Build NerpyBot image
+      - name: Build ${{ matrix.name }} image
         uses: docker/build-push-action@v7
         with:
           context: .
-          target: bot
+          target: ${{ matrix.target }}
           load: true
-          tags: nerpybot:smoke-test
+          tags: ${{ matrix.tag }}
           cache-from: type=gha
           cache-to: type=gha,mode=max
 
       - name: Smoke test NerpyBot
+        if: matrix.target == 'bot'
         run: |
           docker run --rm nerpybot:smoke-test python -c "from bot import NerpyBot; print('OK')"
 
-  validate-migrations:
-    name: Validate Migrations Image
-    runs-on: ubuntu-latest
-    needs: [changes]
-    if: needs.changes.outputs.migrations == 'true'
-    permissions:
-      contents: read
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v6
-        with:
-          fetch-depth: 0
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v4
-
-      - name: Build Migrations image
-        uses: docker/build-push-action@v7
-        with:
-          context: .
-          target: migrations
-          load: true
-          tags: nerpybot-migrations:smoke-test
-          cache-from: type=gha
-          cache-to: type=gha,mode=max
-
-      - name: Smoke test migrations
-        run: |
-          docker run --rm nerpybot-migrations:smoke-test alembic heads
-
-  validate-web:
-    name: Validate Web Dashboard Image
-    runs-on: ubuntu-latest
-    needs: [changes]
-    if: needs.changes.outputs.web == 'true'
-    permissions:
-      contents: read
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v6
-        with:
-          fetch-depth: 0
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v4
-
-      - name: Build Web Dashboard image
-        uses: docker/build-push-action@v7
-        with:
-          context: .
-          target: web
-          load: true
-          tags: nerpybot-web:smoke-test
-          cache-from: type=gha
-          cache-to: type=gha,mode=max
-
       - name: Smoke test Web Dashboard
+        if: matrix.target == 'web'
         run: |
           docker run --rm \
             -e NERPYBOT_WEB_CLIENT_ID=smoke \
