--- .github/workflows/lint.yml
@@ -1,25 +1,0 @@
-name: Lint & Format
-
-on:
-  pull_request:
-    branches: [main]
-
-jobs:
-  lint:
-    name: Ruff Lint & Format Check
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-
-    steps:
-      - uses: actions/checkout@v6
-
-      - name: Run Ruff linter
-        uses: astral-sh/ruff-action@v3
-        with:
-          args: "check --output-format=github"
-
-      - name: Run Ruff formatter check
-        uses: astral-sh/ruff-action@v3
-        with:
-          args: "format --check --diff"
